pyproject.toml
[project]
name = "amap-mcp"
version = "0.1.0"
description = "高德地图 MCP 服务器 - 提供地理编码、路线规划、POI搜索等功能"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorlog>=6.10.1",
    "mcp>=1.22.0",
    "httpx>=0.24.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
